eu.learnpad.lp-cp-apis dependency added
