+++ lp-content-analysis/pom.xml
@@ -38,7 +38,12 @@
 			<artifactId>gate-core</artifactId>
 			<version>8.1</version>
 		</dependency>
-
+		<dependency>
+            <groupId>eu.learnpad</groupId>
+            <artifactId>lp-cp-apis</artifactId>
+            <version>1.0-SNAPSHOT</version>
+           
+        </dependency>
 
 		<dependency>
 			<groupId>org.jsoup</groupId>
