+++ lp-core-platform/lp-cp-apis/src/main/java/eu/learnpad/ca/rest/data/Node.java
@@ -2,10 +2,14 @@ package eu.learnpad.ca.rest.data;
 
 import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlRootElement;
+import javax.xml.bind.annotation.XmlTransient;
 
 
 @XmlRootElement(name = "Node")
-public class Node {
+public class Node implements Comparable<Object>{
+	
+	@XmlTransient
+	protected Integer OffSet;
 
 	protected Integer id=-1;
 	
@@ -19,6 +23,19 @@ public class Node {
 
 	
 	
+	public Node(Integer id,int offSet) {
+		OffSet = offSet;
+		this.id = id;
+	}
+
+	public Integer getOffSet() {
+		return OffSet;
+	}
+
+	public void setOffSet(int offSet) {
+		OffSet = offSet;
+	}
+
 	/**
 	 * Get the value of id.
 	 * 
@@ -64,6 +81,11 @@ public class Node {
 		} else if (!id.equals(other.id))
 			return false;
 		return true;
+	}
+
+	public int compareTo(Object o) {
+		Node node = (Node) o;
+		return OffSet.compareTo(node.getOffSet());
 	}
