+++ lp-core-platform/lp-cp-apis/src/main/java/eu/learnpad/ca/rest/data/collaborative/AnnotatedCollaborativeContentAnalysis.java
@@ -249,6 +249,17 @@ public class AnnotatedCollaborativeContentAnalysis {
 				+  "]\n";
 	}
 
+
+
+	public void setAnnotations(List<Annotation> value) {
+		if (annotations == null) {
+            annotations = new ArrayList<Annotation>();
+            
+        }
+        annotations.addAll(value);
+		
+	}
+
     
 
 }
