+++ lp-core-platform/lp-cp-apis/src/main/java/eu/learnpad/ca/rest/data/stat/AnnotatedStaticContentAnalysis.java
@@ -246,6 +246,16 @@ public class AnnotatedStaticContentAnalysis {
 				+  "]\n";
 	}
 
+
+	public void setAnnotations(List<Annotation> value) {
+		if (annotations == null) {
+            annotations = new ArrayList<Annotation>();
+            
+        }
+        annotations.addAll(value);
+		
+	}
+
     
 
 }
